requirements.txt
sentence-transformers
anthropic
python-dotenv
gradio